refactor(scheduler): centralize recurrence timing
