+++ ARCHITECTURE.md
@@ -1074,7 +1074,8 @@ src/
 │   │   └── store.rs
 │   ├── scheduled_messages/
 │   │   ├── evidence.rs
-│   │   └── postgres_tests.rs
+│   │   ├── postgres_tests.rs
+│   │   └── timing.rs
 │   ├── session_backend/
 │   │   ├── stream_line.rs
 │   │   └── terminal_usage.rs
@@ -1147,6 +1148,7 @@ src/
 │   ├── retrospectives.rs
 │   ├── review_decision.rs
 │   ├── scheduled_messages.rs
+│   ├── scheduling.rs
 │   ├── service_error.rs
 │   ├── session_activity.rs
 │   ├── session_backend.rs
