+++ src/services/mod.rs
@@ -102,6 +102,7 @@ pub mod retrospectives;
 pub mod review_decision;
 pub mod routines;
 pub mod scheduled_messages;
+pub(crate) mod scheduling;
 pub mod service_error;
 pub mod session_activity;
 pub(crate) mod session_selector_validity;
